@@ -34,33 +34,32 @@ dependencies:
   cloud_firestore_web: 2.6.8
   dash_chat: 1.1.16
   equatable: 2.0.3
-  expandable: 5.0.1
+  file_picker: 4.5.1
   firebase_auth: 3.2.0
   firebase_core: 1.12.0
+  firebase_core_platform_interface: 4.4.1
   firebase_core_web: 1.5.4
   firebase_storage: 10.1.0
   firebase_storage_web: 3.2.8
-  flutter_datetime_picker: 1.5.0
-  flutter_google_places: 0.3.0
-  flutter_spinkit: 4.1.2+1
-  font_awesome_flutter: 8.11.0
-  google_fonts: 2.1.0
+  flutter_datetime_picker: 1.5.1
+  flutter_spinkit: 5.1.0
+  font_awesome_flutter: 10.1.0
+  google_fonts: 2.2.0
   google_sign_in: 5.2.1
   image_picker: 0.8.4+4
+  infinite_scroll_pagination: 3.1.0
   intl: 0.17.0
-  json_path: 0.2.1
+  json_path: 0.3.1
   json_serializable: 6.1.3
   mime_type: 1.0.0
   page_transition: 2.0.4
   rxdart: 0.26.0
-  share_plus: ^2.1.4
   shared_preferences: 2.0.11
   sign_in_with_apple:
     git:
       url: https://github.com/asmengistu/dart_packages
       ref: ab18d53a80e0a4bb90f1ad84f37f1b408930f28d
       path: packages/sign_in_with_apple/sign_in_with_apple/
-  smooth_page_indicator: 1.0.0+2
   timeago: 3.1.0
   url_launcher: 6.0.15
 
