@@ -32,7 +32,7 @@ class FFButtonOptions {
   final double iconSize;
   final Color iconColor;
   final EdgeInsetsGeometry iconPadding;
-  final double borderRadius;
+  final BorderRadius borderRadius;
   final BorderSide borderSide;
 }
 
@@ -90,19 +90,55 @@ class _FFButtonWidgetState extends State<FFButtonWidget> {
             setState(() => loading = true);
             try {
               await widget.onPressed();
-            } catch (e) {
-              print('On pressed error:\n$e');
+            } finally {
+              if (mounted) {
+                setState(() => loading = false);
+              }
             }
-            setState(() => loading = false);
           }
         : () => widget.onPressed();
 
+    ButtonStyle style = ButtonStyle(
+      shape: MaterialStateProperty.all<OutlinedBorder>(
+        RoundedRectangleBorder(
+          borderRadius:
+              widget.options.borderRadius ?? BorderRadius.circular(8.0),
+          side: widget.options.borderSide ?? BorderSide.none,
+        ),
+      ),
+      foregroundColor: MaterialStateProperty.resolveWith<Color>(
+        (states) {
+          if (states.contains(MaterialState.disabled)) {
+            return widget.options.disabledTextColor;
+          }
+          return widget.options.textStyle.color;
+        },
+      ),
+      backgroundColor: MaterialStateProperty.resolveWith<Color>(
+        (states) {
+          if (states.contains(MaterialState.disabled)) {
+            return widget.options.disabledColor;
+          }
+          return widget.options.color;
+        },
+      ),
+      overlayColor: MaterialStateProperty.resolveWith<Color>((states) {
+        if (states.contains(MaterialState.pressed)) {
+          return widget.options.splashColor;
+        }
+        return null;
+      }),
+      padding: MaterialStateProperty.all(widget.options.padding ??
+          const EdgeInsets.symmetric(horizontal: 8.0, vertical: 4.0)),
+      elevation:
+          MaterialStateProperty.all<double>(widget.options.elevation ?? 2.0),
+    );
+
     if (widget.icon != null || widget.iconData != null) {
-      textWidget = Flexible(child: textWidget);
       return Container(
         height: widget.options.height,
         width: widget.options.width,
-        child: RaisedButton.icon(
+        child: ElevatedButton.icon(
           icon: Padding(
             padding: widget.options.iconPadding ?? EdgeInsets.zero,
             child: widget.icon ??
@@ -115,18 +151,7 @@ class _FFButtonWidgetState extends State<FFButtonWidget> {
           ),
           label: textWidget,
           onPressed: onPressed,
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(widget.options.borderRadius),
-            side: widget.options.borderSide ?? BorderSide.none,
-          ),
-          color: widget.options.color,
-          colorBrightness:
-              ThemeData.estimateBrightnessForColor(widget.options.color),
-          textColor: widget.options.textStyle.color,
-          disabledColor: widget.options.disabledColor,
-          disabledTextColor: widget.options.disabledTextColor,
-          elevation: widget.options.elevation,
-          splashColor: widget.options.splashColor,
+          style: style,
         ),
       );
     }
@@ -134,21 +159,9 @@ class _FFButtonWidgetState extends State<FFButtonWidget> {
     return Container(
       height: widget.options.height,
       width: widget.options.width,
-      child: RaisedButton(
+      child: ElevatedButton(
         onPressed: onPressed,
-        shape: RoundedRectangleBorder(
-          borderRadius:
-              BorderRadius.circular(widget.options.borderRadius ?? 28),
-          side: widget.options.borderSide ?? BorderSide.none,
-        ),
-        textColor: widget.options.textStyle.color,
-        color: widget.options.color,
-        colorBrightness:
-            ThemeData.estimateBrightnessForColor(widget.options.color),
-        disabledColor: widget.options.disabledColor,
-        disabledTextColor: widget.options.disabledTextColor,
-        padding: widget.options.padding,
-        elevation: widget.options.elevation,
+        style: style,
         child: textWidget,
       ),
     );
