@@ -1,6 +1,5 @@
 import 'package:built_value/standard_json_plugin.dart';
 
-import 'job_posts_record.dart';
 import 'users_record.dart';
 import 'work_history_record.dart';
 import 'companies_record.dart';
@@ -10,6 +9,9 @@ import 'chats_record.dart';
 import 'chat_messages_record.dart';
 import 'candidate_profile_record.dart';
 import 'project_history_record.dart';
+import 'recruiter_profile_record.dart';
+import 'job_posts_record.dart';
+import 'candidature_workflow_record.dart';
 
 import 'index.dart';
 
@@ -21,7 +23,6 @@ part 'serializers.g.dart';
 const kDocumentReferenceField = 'Document__Reference__Field';
 
 @SerializersFor(const [
-  JobPostsRecord,
   UsersRecord,
   WorkHistoryRecord,
   CompaniesRecord,
@@ -31,6 +32,9 @@ const kDocumentReferenceField = 'Document__Reference__Field';
   ChatMessagesRecord,
   CandidateProfileRecord,
   ProjectHistoryRecord,
+  RecruiterProfileRecord,
+  JobPostsRecord,
+  CandidatureWorkflowRecord,
 ])
 final Serializers serializers = (_$serializers.toBuilder()
       ..add(DocumentReferenceSerializer())
