@@ -432,7 +432,7 @@ class _WorkExpBottomSheetWidgetState extends State<WorkExpBottomSheetWidget> {
                               color: Colors.transparent,
                               width: 1,
                             ),
-                            borderRadius: 8,
+                            borderRadius: BorderRadius.circular(8),
                           ),
                         ),
                       ),
