@@ -482,7 +482,7 @@ class _WorkExpEditWidgetState extends State<WorkExpEditWidget> {
                                   color: Colors.transparent,
                                   width: 1,
                                 ),
-                                borderRadius: 8,
+                                borderRadius: BorderRadius.circular(8),
                               ),
                             ),
                           ),
