@@ -4,7 +4,9 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 abstract class FlutterFlowTheme {
-  static FlutterFlowTheme of(BuildContext context) => LightModeTheme();
+  static FlutterFlowTheme of(BuildContext context) {
+    return LightModeTheme();
+  }
 
   Color primaryColor;
   Color secondaryColor;
@@ -24,60 +26,39 @@ abstract class FlutterFlowTheme {
   Color grayIcon400;
   Color lineColor;
   Color dark500;
+  Color gray200;
+  Color gray600;
+  Color black600;
+  Color tertiary400;
+  Color textColor;
 
-  TextStyle get title1 => GoogleFonts.getFont(
-        'Roboto',
-        color: Color(0xFF0C141D),
-        fontWeight: FontWeight.w600,
-        fontSize: 32,
-      );
-  TextStyle get title2 => GoogleFonts.getFont(
-        'Roboto',
-        color: darkText,
-        fontWeight: FontWeight.w500,
-        fontSize: 26,
-      );
-  TextStyle get title3 => GoogleFonts.getFont(
-        'Roboto',
-        color: darkText,
-        fontWeight: FontWeight.w500,
-        fontSize: 24,
-      );
-  TextStyle get subtitle1 => GoogleFonts.getFont(
-        'Roboto',
-        color: darkText,
-        fontWeight: FontWeight.w500,
-        fontSize: 20,
-      );
-  TextStyle get subtitle2 => GoogleFonts.getFont(
-        'Roboto',
-        color: darkText,
-        fontWeight: FontWeight.normal,
-        fontSize: 16,
-      );
-  TextStyle get bodyText1 => GoogleFonts.getFont(
-        'Roboto',
-        color: darkText,
-        fontWeight: FontWeight.normal,
-        fontSize: 14,
-      );
-  TextStyle get bodyText2 => GoogleFonts.getFont(
-        'Roboto',
-        color: grayIcon,
-        fontWeight: FontWeight.normal,
-        fontSize: 12,
-      );
+  String get title1Family => typography.title1Family;
+  TextStyle get title1 => typography.title1;
+  String get title2Family => typography.title2Family;
+  TextStyle get title2 => typography.title2;
+  String get title3Family => typography.title3Family;
+  TextStyle get title3 => typography.title3;
+  String get subtitle1Family => typography.subtitle1Family;
+  TextStyle get subtitle1 => typography.subtitle1;
+  String get subtitle2Family => typography.subtitle2Family;
+  TextStyle get subtitle2 => typography.subtitle2;
+  String get bodyText1Family => typography.bodyText1Family;
+  TextStyle get bodyText1 => typography.bodyText1;
+  String get bodyText2Family => typography.bodyText2Family;
+  TextStyle get bodyText2 => typography.bodyText2;
+
+  Typography get typography => ThemeTypography(this);
 }
 
 class LightModeTheme extends FlutterFlowTheme {
   Color primaryColor = const Color(0xFF20368F);
   Color secondaryColor = const Color(0xFF829CD0);
   Color tertiaryColor = const Color(0xFFFFFFFF);
-  Color alternate = const Color(0x00000000);
-  Color primaryBackground = const Color(0x00000000);
-  Color secondaryBackground = const Color(0x00000000);
-  Color primaryText = const Color(0x00000000);
-  Color secondaryText = const Color(0x00000000);
+  Color alternate = const Color(0xFF39D2C0);
+  Color primaryBackground = const Color(0xFFF1F4F8);
+  Color secondaryBackground = const Color(0xFFFFFFFF);
+  Color primaryText = const Color(0xFF090F13);
+  Color secondaryText = const Color(0xFF95A1AC);
 
   Color grayIcon = Color(0xFF57636C);
   Color darkText = Color(0xFF0C141D);
@@ -88,6 +69,84 @@ class LightModeTheme extends FlutterFlowTheme {
   Color grayIcon400 = Color(0xFFABB3BA);
   Color lineColor = Color(0xFFDBE2E7);
   Color dark500 = Color(0xFF1D2429);
+  Color gray200 = Color(0xFFDBE2E7);
+  Color gray600 = Color(0xFF262D34);
+  Color black600 = Color(0xFF090F13);
+  Color tertiary400 = Color(0xFF39D2C0);
+  Color textColor = Color(0xFF1E2429);
+}
+
+abstract class Typography {
+  String get title1Family;
+  TextStyle get title1;
+  String get title2Family;
+  TextStyle get title2;
+  String get title3Family;
+  TextStyle get title3;
+  String get subtitle1Family;
+  TextStyle get subtitle1;
+  String get subtitle2Family;
+  TextStyle get subtitle2;
+  String get bodyText1Family;
+  TextStyle get bodyText1;
+  String get bodyText2Family;
+  TextStyle get bodyText2;
+}
+
+class ThemeTypography extends Typography {
+  ThemeTypography(this.theme);
+
+  final FlutterFlowTheme theme;
+
+  String get title1Family => 'Roboto';
+  TextStyle get title1 => GoogleFonts.getFont(
+        'Roboto',
+        color: Color(0xFF0C141D),
+        fontWeight: FontWeight.w600,
+        fontSize: 32,
+      );
+  String get title2Family => 'Roboto';
+  TextStyle get title2 => GoogleFonts.getFont(
+        'Roboto',
+        color: theme.darkText,
+        fontWeight: FontWeight.w500,
+        fontSize: 26,
+      );
+  String get title3Family => 'Roboto';
+  TextStyle get title3 => GoogleFonts.getFont(
+        'Roboto',
+        color: theme.darkText,
+        fontWeight: FontWeight.w500,
+        fontSize: 24,
+      );
+  String get subtitle1Family => 'Roboto';
+  TextStyle get subtitle1 => GoogleFonts.getFont(
+        'Roboto',
+        color: theme.darkText,
+        fontWeight: FontWeight.w500,
+        fontSize: 20,
+      );
+  String get subtitle2Family => 'Roboto';
+  TextStyle get subtitle2 => GoogleFonts.getFont(
+        'Roboto',
+        color: theme.darkText,
+        fontWeight: FontWeight.normal,
+        fontSize: 16,
+      );
+  String get bodyText1Family => 'Roboto';
+  TextStyle get bodyText1 => GoogleFonts.getFont(
+        'Roboto',
+        color: theme.darkText,
+        fontWeight: FontWeight.normal,
+        fontSize: 14,
+      );
+  String get bodyText2Family => 'Roboto';
+  TextStyle get bodyText2 => GoogleFonts.getFont(
+        'Roboto',
+        color: theme.grayIcon,
+        fontWeight: FontWeight.normal,
+        fontSize: 12,
+      );
 }
 
 extension TextStyleHelper on TextStyle {
@@ -96,8 +155,10 @@ extension TextStyleHelper on TextStyle {
     Color color,
     double fontSize,
     FontWeight fontWeight,
+    double letterSpacing,
     FontStyle fontStyle,
     bool useGoogleFonts = true,
+    TextDecoration decoration,
     double lineHeight,
   }) =>
       useGoogleFonts
@@ -105,16 +166,20 @@ extension TextStyleHelper on TextStyle {
               fontFamily,
               color: color ?? this.color,
               fontSize: fontSize ?? this.fontSize,
+              letterSpacing: letterSpacing ?? this.letterSpacing,
               fontWeight: fontWeight ?? this.fontWeight,
               fontStyle: fontStyle ?? this.fontStyle,
+              decoration: decoration,
               height: lineHeight,
             )
           : copyWith(
               fontFamily: fontFamily,
               color: color,
               fontSize: fontSize,
+              letterSpacing: letterSpacing,
               fontWeight: fontWeight,
               fontStyle: fontStyle,
+              decoration: decoration,
               height: lineHeight,
             );
 }
