@@ -142,7 +142,7 @@ class _ForgotPasswordWidgetState extends State<ForgotPasswordWidget> {
                   color: Colors.transparent,
                   width: 1,
                 ),
-                borderRadius: 8,
+                borderRadius: BorderRadius.circular(8),
               ),
             ),
           ),
